Commit mess

+++ nodemcu-httpserver-master/http/upload.lua
@@ -1,4 +1,9 @@
+function starts(String,Start)
+  return string.sub(String,1,string.len(Start)) == Start
+end
+
 return function (connection, req, args)
+    collectgarbage()
     dofile("httpserver-header.lc")(connection, 200, 'json')
 	connection:send('{')
 
@@ -14,46 +19,38 @@ return function (connection, req, args)
 	local currentByte = nil
 	
     for name, value in pairs(args) do
+    	fieldsCount = fieldsCount + 1
         if (name == "offset") then
 			mbOffset = tonumber(value, 10)
-
-			fieldsCount = fieldsCount + 1
-		end
-		if (name == "len") then
+		elseif (name == "len") then
 			mbLen = tonumber(value, 10)
-			
-			fieldsCount = fieldsCount + 1
-		end
-		if (name == "data") then
+		elseif (name == "data") then
 			mbData = value
-			
-			fieldsCount = fieldsCount + 1
-		end
-		if (name == "filename") then
+		elseif (name == "filename") then
 			mbFilename = value
-			
-			fieldsCount = fieldsCount + 1
-		end
-		if (name == "filesize") then
+		elseif (name == "filesize") then
 			fileSize = tonumber(value, 10)
-			
-			fieldsCount = fieldsCount + 1
-		end
-		if (name == "cmd") then
+		elseif (name == "cmd") then
 			mbCmd = value
-			
-			fieldsCount = fieldsCount + 1
+		else
+			fieldsCount = fieldsCount - 1
 		end
     end
 	
 	if (mbCmd == 'upload') then
 		if (fieldsCount > 5) then
-			if (mbFilename ~= 'upload.lua') then
+			local outfile = mbFilename
+			if string.len(mbFilename) > 11 then
+			    outfile = 'http.' .. string.sub(mbFilename,0,8) .. string.sub(mbFilename,-4,-1)
+			else
+				outfile = 'http.' .. mbFilename
+			end
+			if (mbFilename ~= 'upload.lc') then
 				connection:send('"offset":"' .. mbOffset .. '",')
 				connection:send('"len":"' .. mbLen .. '",')
 				connection:send('"filename":"' .. mbFilename .. '"')
 				
-				mbFilename = 'http/' .. mbFilename
+
 				
 				for i=1,string.len(mbData),2 do
 					currentByte = tonumber(string.sub(mbData, i, i + 1), 16)
@@ -61,10 +58,9 @@ return function (connection, req, args)
 				end
 				
 				if (mbOffset > 0) then
-					file.open(mbFilename .. '.dnl','a+')
-				else
-					file.remove(mbFilename .. '.dnl')					
-					file.open(mbFilename .. '.dnl','w+')
+					file.open(outfile .. '.dnl','a+')
+				else			
+					file.open(outfile .. '.dnl','w+')
 				end
 				file.seek("set", mbOffset)
 				file.write(binaryData)				
@@ -73,14 +69,28 @@ return function (connection, req, args)
 				binaryData = nil
 				
 				if (fileSize == mbLen + mbOffset) then
-					file.remove(mbFilename)					
-					file.rename(mbFilename .. '.dnl', mbFilename)
-					file.remove(mbFilename .. '.dnl')						
+						
 				
-					if (string.sub(mbFilename, -4) == '.lua') then
-						file.remove(string.sub(mbFilename, 0, -3) .. "lc")
-						node.compile(mbFilename)
-						file.remove(mbFilename)
+					if (mbFilename == 'upload.lua') then
+						-- nothing
+					elseif (mbFilename == 'upload.lc') then
+						-- nothing
+					elseif (mbFilename == 'index.html') then
+						-- nothing
+					elseif (mbFilename == 'index.html.gz') then
+						-- nothing
+					elseif (string.sub(mbFilename, -4) == '.lua' ) then
+						-- file.remove(string.sub(mbFilename, 0, -3) .. "lc")
+						-- node.compile(mbFilename)
+						-- file.remove(mbFilename)
+					elseif (string.sub(mbFilename, -4) == '.lc') then
+						-- file.remove(string.sub(mbFilename, 0, -3) .. "lc")
+						-- node.compile(mbFilename)
+						-- file.remove(mbFilename)
+					else
+						--file.remove(mbFilename)					
+					    file.rename(outfile .. '.dnl', outfile)
+					    file.remove(outfile .. '.dnl')
 					end
 				end		
 			end
@@ -93,18 +103,27 @@ return function (connection, req, args)
 		connection:send('"files":{')
 		
 		for name, size in pairs(file.list()) do
-			local isHttpFile = string.match(name, "(http/)") ~= nil
 			
-			if isHttpFile then
-				if (headerExist > 0) then 
-					connection:send(',')
+			if starts(name, "http.") then
+				local url = string.sub(name, 6, string.len(name))
+
+				if (url == 'upload.lua') then
+					-- nothing
+				elseif (url == 'upload.lc') then
+					-- nothing
+				elseif (url == 'index.html') then
+					-- nothing
+				elseif (url == 'index.html.gz') then
+					-- nothing
+				else
+					if (headerExist > 0) then 
+						connection:send(',')
+					end
+										
+					connection:send('"' .. url .. '":"' .. size .. '"')
+				
+					headerExist = 1
 				end
-				
-				local url = string.match(name, ".*/(.*)")
-				
-				connection:send('"' .. url .. '":"' .. size .. '"')
-				
-				headerExist = 1
 			end
 		end
 		
@@ -113,15 +132,8 @@ return function (connection, req, args)
 		connection:send('"total":"' .. total .. '",')
 		connection:send('"used":"' .. used .. '",')
 		connection:send('"free":"' .. remaining .. '"')
-	elseif (mbCmd == 'remove') then
-		if (fieldsCount > 1) then
-			if (mbFilename ~= 'upload.lua') and (mbFilename ~= 'upload.lc') and (mbFilename ~= 'upload.html.gz') then
-				file.remove('http/' .. mbFilename)
-			end
-		end
 	end
 	
 	connection:send('}')	
 	collectgarbage()
 end
-
